feat: implement useHideAdProps hook for dynamic ad visibility and refactor Sponsor components for improved accessibility
